OnlineExam/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
confusable-homoglyphs==3.2.0
cryptography==3.4.8
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.7
django-allauth==0.45.0
django-ckeditor==6.1.0
django-cors-headers==3.14.0
django-crispy-forms==1.12.0
django-filer==2.2.4
django-import-export==3.1.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-modeladmin-reorder==0.3.1
django-mptt==0.14.0
django-polymorphic==3.1.0
django-registration==3.2
django-rest-auth==0.9.5
django-webpack-loader==1.3.0
djangorestframework==3.12.4
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
idna==3.2
lxml==4.9.2
MarkupPy==1.14
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.4.0
psycopg-binary==3.1.8
psycopg2-binary==2.9.5
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
reportlab==3.6.12
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
svglib==1.5.1
tablib==3.4.0
tinycss2==1.2.1
toml==0.10.2
Unidecode==1.1.2
urllib3==1.26.6
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
